fix DDLGuard lock order in RESTART REPLICAS
