+++ dbms/src/Interpreters/InterpreterRenameQuery.cpp
@@ -54,23 +54,8 @@ BlockIO InterpreterRenameQuery::execute()
     std::vector<RenameDescription> descriptions;
     descriptions.reserve(rename.elements.size());
 
-    /// To avoid deadlocks, we must acquire locks for tables in same order in any different RENAMES.
-    struct UniqueTableName
-    {
-        String database_name;
-        String table_name;
-
-        UniqueTableName(const String & database_name_, const String & table_name_)
-            : database_name(database_name_), table_name(table_name_) {}
-
-        bool operator< (const UniqueTableName & rhs) const
-        {
-            return std::tie(database_name, table_name) < std::tie(rhs.database_name, rhs.table_name);
-        }
-    };
-
     /// Don't allow to drop tables (that we are renaming); don't allow to create tables in places where tables will be renamed.
-    std::map<UniqueTableName, std::unique_ptr<DDLGuard>> table_guards;
+    TableGuards table_guards;
 
     for (const auto & elem : rename.elements)
     {
