+++ dbms/tests/queries/0_stateless/01108_restart_replicas_rename_deadlock.reference
@@ -1,0 +1,1 @@
+1180	40
